requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
pydantic
streamlit
langchain_community